Move things around 9 / 12
Move History to a separate file
