+++ amm/src/main/scala/ammonite/interp/Storage.scala
@@ -253,38 +253,3 @@ object Storage{
     }
   }
 }
-
-class History(s: Vector[String])
-extends IndexedSeq[String]
-with IndexedSeqLike[String, History] {
-  def length: Int = s.length
-  def apply(idx: Int): String = s.apply(idx)
-  override def newBuilder = History.builder
-}
-
-object History{
-  def builder = new mutable.Builder[String, History] {
-    val buffer = mutable.Buffer.empty[String]
-    def +=(elem: String): this.type = {buffer += elem; this}
-
-    def result(): History = new History(buffer.toVector)
-
-    def clear(): Unit = buffer.clear()
-  }
-  implicit def cbf = new CanBuildFrom[History, String, History]{
-    def apply(from: History) = builder
-    def apply() = builder
-  }
-  implicit def toHistory(s: Seq[String]): History = new History(s.toVector)
-
-  import pprint._
-  implicit val historyPPrint: pprint.PPrint[History] = pprint.PPrint(
-    new pprint.PPrinter[History]{
-      def render0(t: History, c: pprint.Config) = {
-        t.iterator.flatMap(Iterator("\n", _))
-      }
-    }
-  )
-
-}
-
